requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
django-extensions==3.2.3
django-uuslug==2.0.0
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.9
python-slugify==8.0.4
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.12.2
